open signals modal on click
