+++ src/components/SignalCard/card/TrendingWordsSignalCard.module.scss
@@ -12,7 +12,6 @@
 
 .header {
   display: flex;
-  margin-top: 4px;
 }
 
 .ampm {
